% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertion.R
\name{as.geonet}
\alias{as.geonet}
\title{Methods for Linear Networks}
\usage{
as.geonet(object)
}
\arguments{
\item{object}{an object of the spatstat class linnet or an object that can
be converted to an instance of this class}
}
\value{
an object of class geonet
}
\description{
The function as.geonet converts an object of class linnet to an object of
class geonet
}
