% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.gn.gnpp}
\alias{as.gn.gnpp}
\title{Methods for Point Patterns on Linear Networks}
\usage{
\method{as.gn}{gnpp}(x, ...)
}
\arguments{
\item{x}{an object of the spatstat class linnet or an object that can
be converted to an instance of this class}

\item{...}{asdd}
}
\value{
Point pattern on a geometric network (object of class \code{gnpp}).
}
\description{
The function as.gnpp.lpp converts an object of class linnet to an object of
class gn
}
