% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.gn.linnet}
\alias{as.gn.linnet}
\title{Methods for Linear Networks}
\usage{
\method{as.gn}{linnet}(x)
}
\arguments{
\item{object}{an object of the spatstat class linnet or an object that can
be converted to an instance of this class}
}
\value{
an object of class gn
}
\description{
The function as.gn converts an object of class linnet to an object of
class gn
}
