% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.gnpp.lpp}
\alias{as.gnpp.lpp}
\title{Methods for Point Patterns on Linear Networks}
\usage{
\method{as.gnpp}{lpp}(x, ...)
}
\arguments{
\item{object}{an object of the spatstat class linnet or an object that can
be converted to an instance of this class}
}
\value{
an object of class gn
}
\description{
The function as.gnpp.lpp converts an object of class linnet to an object of
class gn
}
