% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.linnet.gn}
\alias{as.linnet.gn}
\title{Coerce to Linear Network}
\usage{
\method{as.linnet}{gn}(X, ...)
}
\arguments{
\item{X}{An object of class gn.}
}
\description{
\code{as.linnet.gn} coerces an existing object into a linear network, an object
of class \code{linnet}.
}
\examples{
library(spatstat.data)
library(spatstat.linnet)

x <- as.linnet(small_gn)
plot(x)

L <- simplenet
X <- as_gn(L, spatstat = TRUE)
x <- as.linnet(X)
# TRUE
all.equal(x, L)
}
