% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_gn}
\alias{as_gn}
\alias{as_gn.linnet}
\alias{as_gn.gnpp}
\alias{as_gn.gnppfit}
\title{Coerce to Geometric Network}
\usage{
as_gn(x, ...)

\method{as_gn}{linnet}(x, ..., spatstat = FALSE)

\method{as_gn}{gnpp}(x, ...)

\method{as_gn}{gnppfit}(x, ...)
}
\arguments{
\item{x}{An object that could reasonably be coerced to a geometric network
(object of class gn).}

\item{...}{Further arguments passed to \code{as_gn}.}

\item{spatstat}{Set to \code{TRUE} if retransformation to an object of any
\code{spatstat} class is desired. Some elements of these objects (such as
the window) are discared when being coerced to an object of class
\code{gn}.}
}
\description{
\code{as_gn} coerces an existing object into a geometric network, an object
of class \code{gn}.
}
