% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{bin_data}
\alias{bin_data}
\title{Bin Point Pattern on a Geometric Network}
\usage{
bin_data(X, P, vars, intern)
}
\arguments{
\item{X}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{P}{as}

\item{vars}{as}

\item{intern}{as}
}
\value{
The binned data.
}
\description{
\code{bin_data} bins the data on the supplied point pattern according to
all possible combination of covariates.
}
