% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{binData}
\alias{binData}
\title{Bin Point Pattern on a Geometric Network}
\usage{
binData(X, P, lins, smooths)
}
\arguments{
\item{X}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{P}{as}

\item{lins}{as}

\item{smooths}{as}
}
\value{
The binned data.
}
\description{
\code{binData} performs the actual model fit.
}
