% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{bspline_design_plot}
\alias{bspline_design_plot}
\title{B-Spline Design Matrix for Plotting}
\usage{
bspline_design_plot(X, df)
}
\arguments{
\item{X}{A point pattern on a geometric network (object of class \code{gnpp}).}

\item{df}{A data frame with points at which the fitted intensity should be
plotted.}
}
\value{
A sparse design matrix.
}
\description{
B-Spline Design Matrix for Plotting
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
\keyword{internal}
