% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ends}
\alias{ends}
\title{Methods for Geometric Networks}
\usage{
ends(x)
}
\arguments{
\item{x}{an object of class gn or an object of class linnet (or an object that can
be converted to an instance of class linnet)}
}
\value{
an object of class gn
}
\description{
The function as.gns converts an object of class gn to an object of
class gns
}
