% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitModel}
\alias{fitModel}
\title{Fit a Penalized Spline Poisson Model on a Geometric Network}
\usage{
fitModel(
  design,
  lins = NULL,
  smooths = NULL,
  offset = NULL,
  rho = 10,
  rho_max = 1e+05,
  eps_rho = 0.01,
  maxit_rho = 100
)
}
\arguments{
\item{design}{as}

\item{lins}{as}

\item{smooths}{as}

\item{offset}{as}

\item{rho}{as}

\item{rho_max}{as}

\item{eps_rho}{as}

\item{maxit_rho}{as}
}
\value{
Model fit.
}
\description{
\code{fitModel} performs the actual model fit.
}
