% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonet.R
\name{geonet}
\alias{geonet}
\title{Geonet Map Projection}
\usage{
geonet(g, d, mode = 'globejs', filter = 'off', ..., options = geonetOptions())
}
\arguments{
\item{g}{Undirected network, iGraph.}

\item{d}{Data frame containing meta data for each vertex (ordered as vertices in graph); requires columns 'ID', 'Longitude', 'Latitude', 'Colour', 'Size'.}

\item{mode}{Mapping mode, one of: 'leaflet', 'globejs', 'sample'. [ string, 'globejs' ]}

\item{filter}{Enable colour filter for edges, one of: 'different', 'same', 'off'. [ string, 'off' ]}

\item{time}{Decompose graph and map subgraphs by time points. [ bool, FALSE ]}

\item{shiny}{Show results in simple Shiny App. [ bool, FALSE ]}

\item{save}{Save visualization as project in current working directory. [ bool, FALSE ]}

\item{project}{Project name for saving output. [ string, 'project' ]}

\item{options}{Options from geonetOptions(). [ list, geonetOptions() ]}
}
\value{
list of map objects
}
\description{
Generate a map projection of an undirected network with Leaflet or globeJS.
}

