% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptions.R
\name{geonetOptions}
\alias{geonetOptions}
\title{Geonet Options}
\usage{
getOptions(jitter = 0, nodeColour = 'Colour', nodeSize = 'Size', nodeFill = nodeColour, ...)
}
\arguments{
\item{jitter}{Randomly displace nodes by given amount of Longitude / Latitude. [ float, 0 ]}

\item{nodeColour}{Name of column containing colours for the halo of each node. [ string, 'Colour' ]}

\item{nodeSize}{Name of column containing sizes for the halo of each node. [ string, 'Size' ]}

\item{nodeFill}{Name of column containing fill colours for each node. [ string, 'nodeColour' ]}

\item{timeData}{Name of column containing time points for each node. [ string, 'Time' ]}

\item{filterValue}{Name of column containing the data to filter edges by. [ string, 'ID' ]}

\item{edgeColours}{Colours for un-filtered and filtered edges. [ vector, c('yellow', 'red') ]}

\item{containHTML}{Switch for writing data as self-contained HTML. [ bool, FALSE ]}

\item{leaflet.mapStyle}{Map style for Leaflet. [ string, 'CartoDB.DarkMatter' ]}

\item{leaflet.infoData}{Names of columns whose data to show in node popups. [ vector, c('ID') ]}

\item{leaflet.lineData}{Edge width and opacity. [ vector, c(1.5, 0.7) ]}

\item{leaflet.mapView}{Initial view given by Longitude, Latitude, Zoom. [ vector, c(0,0,3) ]}

\item{leaflet.nodeOpacity}{Node halo opacity. [ float, 0.5 ]}

\item{leaflet.fillOpacity}{Node fill opacity. [ float, 0.2 ]}

\item{leaflet.timeLayers}{Switch to generate a single map with one layer for each time point subgraph. [ bool, FALSE ]}

\item{leaflet.edgePopup}{Switch to add infoData to edges. [ bool, FALSE ]}

\item{leaflet.nodeLayers}{List of layer headers (names) and associated list (value) specifiying column names for node colour, size and fill. [ list, NULL ]}

\item{leaflet.legend}{Switch to add a legend to map. [ bool, FALSE ]}

\item{leaflet.legendColours}{Colours for each legend entry. [ vector, c('red', 'green', 'yellow') ]}

\item{leaflet.legendLabels}{Legend labels. [ vector, c('Label1', 'Label2', 'Label3') ]}

\item{leaflet.legendPosition}{Legend position [ string, 'bottomleft' ]}

\item{globe.arcHeight}{Arc height. [ float, 0.6 ]}

\item{globe.arcWidth}{Arc width. [ float, 2 ]}

\item{globe.arcOpacity}{Arc opacity [ float, 0.5 ]}

\item{globe.atmosphere}{Enable atmosphere. [ bool, FALSE ]}

\item{globe.backgroundColour}{Background colour. [ string, 'black' ]}

\item{globe.bodyColour}{Globe body colour. [ string, '#0000ff' ]}

\item{globe.emissiveColour}{Globe emissive colour. [ string, '#0000ff' ]}

\item{globe.lightColour}{Environmental light colour. [ string, '#aaeeff' ]}

\item{globe.viewField}{Initial field of view. [ int, 35 ]}

\item{globe.rotLatitude}{Initial view rotational Latitude. [ float, 0 ]}

\item{globe.rotLongitude}{Initial view rotational Longitude. [ float, 0 ]}

\item{globe.pointSize}{Bar width [ float, 1 ]}

\item{globe.renderer}{Renderer for globeJS. [ string, 'auto' ]}

\item{globe.surfaceImage}{File path or URL of image to plot on globe surface. [ str, system.file("images/world.jpg", package = "threejs") ]}
}
\value{
list of options for map projection with Leaflet and globeJS
}
\description{
Get options for map projections with Leaflet and globeJS. For details on the effect of these options,
please see the manual or the documentation for Leaflet and globeJS.
}

