% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_confidence_band}
\alias{get_confidence_band}
\title{Intensity Estimation on Geometric Networks with Penalized Splines}
\usage{
get_confidence_band(theta, V, design, i, ind, smooths, q = 0.05, R = 10000)
}
\arguments{
\item{theta}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{V}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{design}{as}

\item{i}{as}

\item{ind}{as}

\item{smooths}{The knot distance delta}

\item{q}{The bin width h}

\item{R}{The order of the penalty}
}
\value{
an object of class gnppfit
}
\description{
\code{intensityPspline} estimates the intensity of a point pattern on a
geometric network.
}
