% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{getB}
\alias{getB}
\title{B-Spline Model Matrix}
\usage{
getB(x)
}
\arguments{
\item{x}{an object of class gn augmented by}
}
\value{
model matrix B
}
\description{
The function calculates the B-spline model matrix B
}
