% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{getBplot}
\alias{getBplot}
\title{B-Spline Design for Plotting}
\usage{
getBplot(X, df)
}
\arguments{
\item{X}{A point pattern on a geometric network (object of class \code{gnpp}).}

\item{df}{A data frame with points at which the fitted intensity should be
plotted.}
}
\value{
A sparse design matrix.
}
\description{
B-Spline Design for Plotting
}
