% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilterStatus.R
\name{getFilterStatus}
\alias{getFilterStatus}
\title{Helper Function: Filter Status}
\usage{
getFilterStatus(src, target, data, filter, filterValue)
}
\description{
Helper function to get filter status for each edge in the network.
}

