% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{getIncidence}
\alias{getIncidence}
\title{Get the incidence matrix}
\usage{
getIncidence(x)
}
\arguments{
\item{x}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}
}
\value{
an object of class gnds
}
\description{
The function as.gnds converts an object of class gnpp to an object of
class gnds
}
