% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{getK}
\alias{getK}
\title{Penalty Matrix K}
\usage{
getK(x, r)
}
\arguments{
\item{x}{an object of class gn augmented by}

\item{r}{the order of the penalty}
}
\value{
penalty matrix K
}
\description{
The function calculates the penalty matrix L
}
