% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjection.R
\name{getProjection}
\alias{getProjection}
\title{Helper Function: Projection}
\usage{
getProjection(data, map.data, mode = "globejs", time = FALSE,
  filter = "off", options = NULL)
}
\description{
Helper function to construct network map projection.
}

