% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectionData.R
\name{getProjectionData}
\alias{getProjectionData}
\title{Helper Function: Projection Data}
\usage{
getProjectionData(graph, data, mode = "globejs", filter = "off",
  options = NULL)
}
\description{
Helper function to transform graph and meta data to map projection format.
}

