% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{incidence}
\alias{incidence}
\title{Incidence matrix of a geometric network}
\usage{
incidence(vertices, lins)
}
\arguments{
\item{vertices}{A data frame containing the vertices of the geometric
network.}

\item{lins}{A data frame containing the curve segments of the geometric
network.}
}
\value{
The incidence matrix of dimension \eqn{W} by \eqn{M}.
}
\description{
\code{incidence} constructs the incidence matrix of a geometric network
from the vertices and the curve segments.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
\keyword{internal}
