% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{intensityPspline}
\alias{intensityPspline}
\title{Intensity Estimation on Geometric Networks with Penalized Splines}
\usage{
intensityPspline(
  X,
  lins = NULL,
  smooths = NULL,
  offset = NULL,
  delta = NULL,
  h = NULL,
  r = 1,
  density = FALSE
)
}
\arguments{
\item{X}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{lins}{as}

\item{smooths}{as}

\item{offset}{as}

\item{delta}{The knot distance delta}

\item{h}{The bin width h}

\item{r}{The order of the penalty}

\item{density}{asd}
}
\value{
an object of class gnppfit
}
\description{
\code{intensityPspline} estimates the intensity of a point pattern on a
geometric network.
}
