% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{internal}
\alias{internal}
\title{Internal Covariates}
\usage{
internal(vars, X, P)
}
\arguments{
\item{vars}{The name of the covariates which should go into the model
as linear internal covariates.}

\item{X}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{P}{Penalized spline representation of the geometric network \code{X}.}
}
\value{
A data frame with the number of rows as the number of bins on the
geometric network (\code{sum(P$bins$N)}) and the number of columns equal to
the length of \code{vars}.
}
\description{
\code{internal} computes the values of internal covariates in every
bin of the network.
}
