% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{logL}
\alias{logL}
\title{Get the incidence matrix}
\usage{
logL(theta, design, rho)
}
\arguments{
\item{theta}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}

\item{design}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}

\item{rho}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}
}
\value{
an object of class gnds
}
\description{
The function as.gnds converts an object of class gnpp to an object of
class gnds
}
