% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{montgomery}
\alias{montgomery}
\title{Car Crashes on Highways in Montgomery County}
\format{
\code{montgomery} is an object of class "\code{gnpp}".
}
\source{
Created by Marc Schneble \email{marc.schneble@lmu.de}.
}
\usage{
data(montgomery)
}
\description{
asd
}
\keyword{datasets}
