% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_bins}
\alias{network_bins}
\title{Defining bins on a Geometric Network}
\usage{
network_bins(G, h = NULL)
}
\arguments{
\item{G}{An object of class \code{gn}) or a point pattern on a geometric
network (object of class \code{gnpp}).}

\item{h}{The global bin width h.}
}
\value{
A list which contains the bins of every curve of the geometric
network.
}
\description{
\code{network_bins} subdivides each curve segment into several bins.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
