% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_intensity_edge}
\alias{network_intensity_edge}
\title{Title}
\usage{
network_intensity_edge(z, m, fit1, fit2 = NULL)
}
\arguments{
\item{z}{a}

\item{m}{a}

\item{fit1}{a}

\item{fit2}{a}
}
\value{

}
\description{
Title
}
