% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_ISE}
\alias{network_ISE}
\title{Computation of the Integrated Squared Error}
\usage{
network_ISE(fit1, fit2)
}
\arguments{
\item{fit1}{The true intensity.}

\item{fit2}{The estimated intensity.}
}
\value{
The (normalized) integrated squared error.
}
\description{
Computes the integrated squared error between a true between a true
intensity \code{fit1} and an estimate \code{fit2}.
}
