% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_knots}
\alias{network_knots}
\title{Defining knots on a Geometric Network}
\usage{
network_knots(G, delta = NULL)
}
\arguments{
\item{G}{An object of class \code{gn}) or a point pattern on a geometric
network (object of class \code{gnpp}).}

\item{delta}{The global knot distance delta.}
}
\value{
A list which contains the knot sequence of every curve of the
geometric network.
}
\description{
\code{network_knots} defines knots on a geometric network (object of class
\code{gn}) which can be used to construct linear B-splines on it.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
