% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_location}
\alias{network_location}
\title{Find Location of a Point on a Geometric Network}
\usage{
network_location(G, m, z)
}
\arguments{
\item{G}{A geometric network.}

\item{m}{The segment index.}

\item{z}{The shortest path distance from the beginning of the network
segment.}
}
\value{
A list with names l, tp_l, x and y.
}
\description{
Find Location of a Point on a Geometric Network
}
