% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{penalty_network}
\alias{penalty_network}
\title{Penalty Matrix of a Geometric Network}
\usage{
penalty_network(G, P, r)
}
\arguments{
\item{G}{A geometric network (object of class \code{gnpp}).}

\item{P}{B-splines on a geometric network.}

\item{r}{The order of the penalty, default to first-order penalty (\code{r = 1}.}
}
\value{
A sparse and square penalty matrix.
}
\description{
\code{penalty_network} constructs the penalty matrix which relates to the B-Spline
design matrix created by \code{\link[geonet]{bspline_design}}.
}
