% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.geonet}
\alias{plot.geonet}
\title{Methods for Geometric Networks}
\usage{
\method{plot}{geonet}(x, ..., x.title = "x", y.title = "y", title = NULL)
}
\arguments{
\item{x}{an object of class \code{geonet}}

\item{...}{further arguments passed to plot}

\item{x.title}{x-axis title}

\item{y.title}{y-axis title}

\item{title}{plot title}
}
\value{
an object of class ggplot
}
\description{
This is the \code{plot} method for an object of class \code{geonet}.
}
\details{
Details
}
