% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gn}
\alias{plot.gn}
\title{Plot Method for Geometric Networks}
\usage{
\method{plot}{gn}(x, ..., title = "", title_x = "", title_y = "", size = 1, frame = TRUE)
}
\arguments{
\item{x}{A geometric network (object of class \code{gn}).}

\item{...}{Further arguments passed to plot.}

\item{title}{Main title of the plot.}

\item{title_x}{x-axis title (ignored if \code{frame = FALSE}).}

\item{title_y}{y-axis title (ignored if \code{frame = FALSE}).}

\item{size}{Plotting size of the line segments. This specificies the
\code{size} argument of \code{\link[ggplot2]{geom_segment}}.}

\item{frame}{Should a frame be drawn around the network?}
}
\value{
Invisibly returns an object of class \code{ggplot}.
}
\description{
This is the \code{plot} method for an object of class \code{gn}.
}
\details{
Details
}
