% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.gnd}
\alias{plot.gnd}
\title{Plot Method for Data on Geometric Networks}
\usage{
\method{plot}{gnd}(x, ..., title_x = "x", title_y = "y", title = "")
}
\arguments{
\item{x}{A geometric network with data (an object of class \code{gns}).}

\item{...}{further arguments passed to plot}

\item{title}{plot title}
}
\value{
an object of class ggplot
}
\description{
This is the \code{plot} method for an object of class \code{gns}.
}
\details{
Details
}
