% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gnppfit}
\alias{plot.gnppfit}
\title{Plot Method for Fitted Models on Geometric Networks}
\usage{
\method{plot}{gnppfit}(
  x,
  ...,
  select = NULL,
  title = "",
  title_x = "x",
  title_y = "y",
  size = 1,
  frame = TRUE,
  sol = 100
)
}
\arguments{
\item{x}{A model fit on a geometric network
(an object of class \code{gnppfit}).}

\item{...}{Further arguments passed to plot}

\item{select}{Allows the plot for a single model term to be selected for
printing. e.g. if you just want the plot for the second smooth term set
\code{select = 2}.}

\item{title}{Main title of the plot.}

\item{title_x}{x-axis title (ignored if \code{frame = FALSE}).}

\item{title_y}{y-axis title (ignored if \code{frame = FALSE}).}

\item{size}{Plotting size of the line segments. This specificies the
\code{size} argument of \code{\link[ggplot2]{geom_segment}}.}

\item{frame}{Should a frame be drawn around the network?}

\item{sol}{asd}
}
\value{
Invisibly returns a list of objects with elements of class ggplot.
}
\description{
This is the \code{plot} method for an object of class \code{gnppfit}.
}
\details{
Details
}
