% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.geonet}
\alias{print.geonet}
\title{Methods for Geometric Networks}
\usage{
\method{print}{geonet}(x, ...)
}
\arguments{
\item{x}{an object of class geonet}

\item{...}{further arguments passed to print}
}
\description{
This is the print method for object of the class geonet.
}
