% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gnpp}
\alias{print.gnpp}
\title{Print Method for Geometric Networks Data}
\usage{
\method{print}{gnpp}(x, ...)
}
\arguments{
\item{x}{an object of class gns}

\item{...}{further arguments passed to print}
}
\description{
This is the print method for object of the class gnd.
}
