% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.gns}
\alias{print.gns}
\title{Print Method for Geometric Networks Splines}
\usage{
\method{print}{gns}(x, ...)
}
\arguments{
\item{x}{an object of class gns}

\item{...}{further arguments passed to print}
}
\description{
This is the print method for object of the class gns.
}
