% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.gn}
\alias{print.summary.gn}
\title{Print Method for Summaries}
\usage{
\method{print}{summary.gn}(x, ...)
}
\arguments{
\item{x}{A geonet summary object of class \code{summary.gn},
code{summary.gnpp} or \code{summary.gnppfit}.}

\item{...}{Other arguments.}
}
\value{
Invisibly returns the NULL object.
}
\description{
Prints basic information of a geonet summary object.
}
