% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.gnppfit}
\alias{print.summary.gnppfit}
\title{Print Method for gnppfit Summary}
\usage{
\method{print}{summary.gnppfit}(x, ...)
}
\arguments{
\item{x}{A model fit on a geometric network (object of class \code{gnpp}).}

\item{...}{further arguments passed to print}
}
\description{
Prints basic information on the model fit and the underlying geometric
network.
}
