% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{pspline}
\alias{pspline}
\title{Penalized Splines on a Geometric Network}
\usage{
pspline(G, delta = NULL, h = NULL)
}
\arguments{
\item{G}{A geometric network (object of class \code{gn}) or,
alternatively, a point pattern on a geometric network
(object of class \code{gnpp}).}

\item{delta}{The knot distance delta.}

\item{h}{The bin width h.}
}
\value{
Linear B-Splines and bins on the geometric network.
}
\description{
\code{pspline} constructs linear B-splines on a geometric network and
subdivides each curve segment into severval bins.
}
