% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runifgn}
\alias{runifgn}
\title{Simluate Uniform Points on a Geometric Network}
\usage{
runifgn(n, G)
}
\arguments{
\item{n}{Number of random points to generate. A nonnegative integer.}

\item{G}{A geometric network (object of class "gn", see gn)}
}
\value{
an object of class gnd
}
\description{
The function runifgn simulates points on a geometric network
}
