% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProject.R
\name{saveProject}
\alias{saveProject}
\title{Helper Function: Save Project}
\usage{
saveProject(graphs, maps, mode = "globejs", project = "geonet",
  time = FALSE, options = NULL)
}
\description{
Helper function to save project HTML.
}

