% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{score}
\alias{score}
\title{Maximum-Likelihood Estimation}
\usage{
score(theta, design, rho)
}
\arguments{
\item{theta}{A vector of model coefficients.}

\item{design}{The design of the model.}

\item{rho}{A vector of smoothing parameters.}
}
\value{
A scalar (\code{logL}), a vector of the same length as
theta (\code{score}) or a square matrix of the same dimension as
theta (\code{fisher}).
}
\description{
\code{logL}, \code{score} and \code{fisher} compute the log-likelihood,
the score-function and the Fisher-Information of the Poisson model.
}
