% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_gn}
\alias{small_gn}
\title{Simple example of a Geometric Network}
\format{
\code{small_gn} is an object of class "\code{gn}".
}
\source{
Created by Marc Schneble \email{marc.schneble@lmu.de}.
}
\usage{
data(small_gn)
}
\description{
A simple, artificially created, example of a geometric network.
}
\keyword{datasets}
