% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/summary.R
\name{print.summary.gnpp}
\alias{print.summary.gnpp}
\alias{print.summary.gnppfit}
\alias{summary.gn}
\alias{summary.gnpp}
\title{Summary for a (point pattern on a) geometric network}
\usage{
\method{print}{summary.gnpp}(x, ...)

\method{print}{summary.gnppfit}(x, ...)

\method{summary}{gn}(object, ...)

\method{summary}{gnpp}(object, ...)
}
\arguments{
\item{...}{Other arguments.}

\item{object}{A geometric network (object of class \code{gn} or a point
pattern on a geometric network (object of class \code{gnpp}).}
}
\value{
A summary of a (point pattern on a) geometric network. This is an
object of class \code{summary.gn} or \code{summary.gnpp}, respectively.
}
\description{
Takes an object of class \code{gn} or \code{gnpp} and computes a summary
from it.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
