% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gnppfit}
\alias{summary.gnppfit}
\title{Intensity Estimation on Geometric Networks with Penalized Splines}
\usage{
\method{summary}{gnppfit}(object, ...)
}
\arguments{
\item{...}{asdad}

\item{x}{Point pattern on a geometric network (object of class \code{gnpp})}
}
\value{
A fitted geometric network (object of class \code{gnppfit}).
}
\description{
\code{intensityPspline} estimates the intensity of a point pattern on a
geometric network.
}
