% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.gns}
\alias{as.gns}
\title{Methods for Geometric Networks}
\usage{
as.gns(x, delta = NULL, h = NULL, r = 1)
}
\arguments{
\item{x}{an object of class gn or an object of class linnet (or an object that can
be converted to an instance of class linnet)}

\item{delta}{The knot distance delta}

\item{h}{The bin width h}

\item{r}{The order of the penalty}
}
\value{
an object of class gn
}
\description{
The function as.gns converts an object of class gn to an object of
class gns
}
