% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_gnpp}
\alias{as_gnpp}
\alias{as_gnpp.gnppfit}
\alias{as_gnpp.lpp}
\title{Coerece to Point Pattern on a Geometric Network}
\usage{
as_gnpp(x, ...)

\method{as_gnpp}{gnppfit}(x, ...)

\method{as_gnpp}{lpp}(x, ..., spatstat = FALSE)
}
\arguments{
\item{x}{An object that could reasonable coerced to a point pattern on a
geometric network (object of class \code{gnpp}).}

\item{...}{further}

\item{spatstat}{Set to \code{TRUE} if retransformation to an object of any
\code{spatstat} class is desired. Some elements of these objects (such as
the window) are discared when being coerced to an object of class
\code{gn}.}
}
\description{
\code{as_gnpp} coerces an existing object into a point pattern on a geometric
network, an object
of class \code{gnpp}.
}
