% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_lpp}
\alias{as_lpp}
\alias{as_lpp.gnpp}
\title{Coerece to Point Pattern on a Linear Network}
\usage{
as_lpp(x, ...)

\method{as_lpp}{gnpp}(x, ...)
}
\arguments{
\item{x}{An object of class gnpp.}

\item{...}{further}
}
\description{
\code{as_lpp} coerces an existing object into a point pattern on a geometric
network, an object
of class \code{lpp}.
}
\examples{
library(spatstat.data)
library(spatstat.linnet)

x <- as_lpp(Montgomery)
plot(x)

L <- simplenet
X <- as_gn(L, spatstat = TRUE)
x <- as.linnet(X)
# TRUE
all.equal(x, L)
}
