% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{bin_data}
\alias{bin_data}
\title{Bin Point Pattern on a Geometric Network}
\usage{
bin_data(X, bins = NULL, vars = NULL, vars_internal = NULL, scale = NULL)
}
\arguments{
\item{X}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{bins}{A list containing the bins of the geometric network.}

\item{vars}{A character vector containing the name of all covariates in the
model.}

\item{vars_internal}{A character vector containing the name of all interval
covariates in the model.}

\item{scale}{A named list which specifies the rescaling of network related
covariates. Currently, internal covariates "x", "y", and "dist2V" can be
scaled.}
}
\value{
The binned data.
}
\description{
\code{bin_data} bins the data on the supplied point pattern according to
all possible combination of covariates.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
