% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{binData}
\alias{binData}
\title{Bin Point Pattern on a Geometric Network}
\usage{
binData(X, P, vars, intern)
}
\arguments{
\item{X}{Point pattern on a geometric network (object of class \code{gnpp})}

\item{P}{as}

\item{lins}{as}

\item{smooths}{as}
}
\value{
The binned data.
}
\description{
\code{binData} bins the data on the supplied point pattern according to
all possible combination of covariates.
}
