% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{bspline_design}
\alias{bspline_design}
\title{Design Matrix of Linear B-Splines on a Geometric Network}
\usage{
bspline_design(G, P)
}
\arguments{
\item{G}{A geometric network (object of class \code{gn}).}

\item{P}{B-splines on a geometric network.}
}
\value{
A sparse matrix.
}
\description{
\code{bspline_design} constructs the design matrix which represents
the baseline intensity on the geometric network.
}
