% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{getBplot}
\alias{getBplot}
\title{Get the incidence matrix}
\usage{
getBplot(X, df)
}
\arguments{
\item{X}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}

\item{df}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}
}
\value{
an object of class gnds
}
\description{
The function as.gnds converts an object of class gnpp to an object of
class gnds
}
