% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_integral}
\alias{network_integral}
\title{Integral of a fitted intensity}
\usage{
network_integral(fit)
}
\arguments{
\item{fit}{A fitted point process on a geometric network.}
}
\value{
The integral.
}
\description{
Integral of a fitted intensity
}
