% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gn}
\alias{plot.gn}
\title{Plot Method for Geometric Networks}
\usage{
\method{plot}{gn}(x, ..., x.title = "x", y.title = "y", title = NULL, size = 1)
}
\arguments{
\item{x}{an object of class \code{gn}}

\item{...}{further arguments passed to plot}

\item{x.title}{x-axis title}

\item{y.title}{y-axis title}

\item{title}{plot title}
}
\value{
an object of class ggplot
}
\description{
This is the \code{plot} method for an object of class \code{gn}.
}
\details{
Details
}
