% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gnpp}
\alias{plot.gnpp}
\title{Plot Method for Points Patterns on Geometric Networks}
\usage{
\method{plot}{gnpp}(
  x,
  ...,
  covariate = NULL,
  title = "",
  title_x = "",
  title_y = "",
  size_lines = 1,
  size_points = 1,
  frame = TRUE
)
}
\arguments{
\item{x}{A point pattern on a geometric network
(object of class \code{gnpp}).}

\item{...}{Further arguments passed to plot.}

\item{covariate}{Character vector of length 1, name of the covariate
that should be plotted.}

\item{title}{Main title of the plot.}

\item{title_x}{x-axis title (ignored if \code{frame = FALSE}).}

\item{title_y}{y-axis title (ignored if \code{frame = FALSE}).}

\item{size_lines}{Plotting size of the line segments. This specificies the
\code{size} argument of \code{\link[ggplot2]{geom_segment}}.}

\item{size_points}{Plotting size of the point pattern. This specificies the
\code{size} argument of \code{\link[ggplot2]{geom_point}}.}

\item{frame}{Should a frame be drawn around the network?}
}
\value{
Invisibly returns an object of class \code{ggplot}.
}
\description{
This is the \code{plot} method for an object of class \code{gnpp}.
}
\details{
Details
}
