% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gnpp}
\alias{plot.gnpp}
\title{Plot Method for Points Patterns on Geometric Networks}
\usage{
\method{plot}{gnpp}(x, ..., title_x = "x", title_y = "y", title = "")
}
\arguments{
\item{x}{A geometric network with data (an object of class \code{gnpp}).}

\item{...}{further arguments passed to plot}

\item{title}{plot title}
}
\value{
an object of class ggplot
}
\description{
This is the \code{plot} method for an object of class \code{gnpp}.
}
\details{
Details
}
