% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gnppfit}
\alias{plot.gnppfit}
\title{Plot Method for Points Patterns on Geometric Networks}
\usage{
\method{plot}{gnppfit}(x, ..., title_x = "x", title_y = "y", title = "", size = 1, sol = 100)
}
\arguments{
\item{x}{A geometric network with data (an object of class \code{gnppfit}).}

\item{...}{further arguments passed to plot}

\item{title_x}{asd}

\item{title_y}{asd}

\item{title}{plot title}

\item{size}{asdas}

\item{sol}{asd}
}
\value{
an object of class ggplot
}
\description{
This is the \code{plot} method for an object of class \code{gnpp}.
}
\details{
Details
}
