% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.gn}
\alias{print.gn}
\title{Methods for Geometric Networks}
\usage{
\method{print}{gn}(x, ...)
}
\arguments{
\item{x}{an object of class gn}

\item{...}{further arguments passed to print}
}
\description{
This is the print method for object of the class gn.
}
