% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gn}
\alias{print.gn}
\alias{print.gnpp}
\alias{print.gnppfit}
\title{Print Method for Geometric Networks}
\usage{
\method{print}{gn}(x, ...)

\method{print}{gnpp}(x, ...)

\method{print}{gnppfit}(x, ...)
}
\arguments{
\item{x}{A geometric network (object of class \code{gn}), a
point pattern on a geometric network (object of class \code{gnpp}) or a
fitted point process on a geometric network (object of class \code{gnppfit}).}

\item{...}{Other arguments.}
}
\value{
Invisibly returns the NULL object.
}
\description{
Prints basic information of a geometric network related object.
}
