% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gn}
\alias{print.gn}
\alias{print.gnpp}
\title{Print Method for Geometric Networks}
\usage{
\method{print}{gn}(x, ...)

\method{print}{gnpp}(x, ...)
}
\arguments{
\item{x}{A geometric network (object of class \code{gn}) or a
point pattern on a geometric network (object of class \code{gnpp}).}

\item{...}{further arguments passed to print}
}
\description{
Prints basic information on the geometric network
}
