% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{Pspline}
\alias{Pspline}
\title{Methods for Geometric Networks}
\usage{
Pspline(X, delta = NULL, h = NULL, r = 1)
}
\arguments{
\item{X}{an object of class gnd or an object of class linnet (or an object that can
be converted to an instance of class linnet)}

\item{delta}{The knot distance delta}

\item{h}{The bin width h}

\item{r}{The order of the penalty}
}
\value{
an object of class gnds
}
\description{
The function as.gnds converts an object of class gnpp to an object of
class gnds
}
