% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rgnpp}
\alias{rgnpp}
\title{Random Points on a Geometric Network}
\usage{
rgnpp(n, fit)
}
\arguments{
\item{n}{Number of random points. A nonnegative integer.}

\item{fit}{A fitted point process on a geometric network (object of class
\code{gnpp}).}
}
\value{
A point pattern on a geometric network, an object of class
\code{gnpp}.
}
\description{
Random Points on a Geometric Network
}
