% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{runifgn}
\alias{runifgn}
\title{Simulate Uniform Points on a Geometric Network}
\usage{
runifgn(n, G)
}
\arguments{
\item{n}{Number of random points to generate. A nonnegative integer.}

\item{G}{A geometric network (object of class \code{gn}).}
}
\value{
An object of class \code{gnpp}.
}
\description{
\code{runifgn} simulates points on a geometric network according
to a uniform density.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
