% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gnppfit}
\alias{summary.gnppfit}
\title{Summary for fitted point process on a geometric network}
\usage{
\method{summary}{gnppfit}(object, ...)
}
\arguments{
\item{object}{Fitted point process on a geometric network (object of class
\code{gnppfit}).}

\item{...}{Other arguments.}
}
\value{
A summary of a fitted point process on a geometric network (object of
class \code{summary.gnppfit}).
}
\description{
Takes a fitted \code{gnppfit} object produced by \code{intensity_spline} and
computes a summary from it.
}
\author{
Marc Schneble \email{marc.schneble@stat.uni-muenchen.de}
}
